+++ src/VIC.DataAccess.MySql/Core/MySqlDataCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Pomelo.Data.MySql;
 using System.Data.Common;
-using VIC.DataAccess.Abstraction;
+using VIC.DataAccess.Abstraction.Converter;
 using VIC.DataAccess.Core;
 
 namespace VIC.DataAccess.MySql.Core
