+++ src/VIC.DataAccess.PostgreSQL/Core/PostgreSQLDataCommand.cs
@@ -1,6 +1,6 @@
 ﻿using Npgsql;
 using System.Data.Common;
-using VIC.DataAccess.Abstraction;
+using VIC.DataAccess.Abstraction.Converter;
 using VIC.DataAccess.Core;
 
 namespace VIC.DataAccess.PostgreSQL.Core
