+++ src/VIC.DataAccess.Sqlite/DataAccessExtensions.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.Data.Sqlite;
 using Microsoft.Extensions.DependencyInjection;
 using VIC.DataAccess.Abstraction;
+using VIC.DataAccess.Abstraction.Converter;
 using VIC.DataAccess.Core;
+using VIC.DataAccess.Core.Converter;
 using VIC.DataAccess.SQLite.Core;
 
 namespace VIC.DataAccess
@@ -11,7 +13,12 @@ namespace VIC.DataAccess
         public static IServiceCollection UseDataAccess(this IServiceCollection service)
         {
             TypeHelper.SqlParameterType = typeof(SqliteParameter);
-            return service.AddTransient<IDataCommand, SQLiteDataCommand>();
+            return service.AddSingleton<IDbFuncNameConverter, DbFuncNameConverter>()
+                .AddSingleton<IDbTypeConverter, DbTypeConverter>()
+                .AddSingleton<IScalarConverter, ScalarConverter>()
+                .AddSingleton<IEntityConverter, EntityConverter>()
+                .AddSingleton<IParamConverter, ParamConverter>()
+                .AddTransient<IDataCommand, SQLiteDataCommand>();
         }
     }
 }
