+++ src/VIC.DataAccess/Abstraction/Converter/IScalarConverter.cs
@@ -1,8 +1,8 @@
 ﻿using System.Data.Common;
 
-namespace VIC.DataAccess.Abstraction
+namespace VIC.DataAccess.Abstraction.Converter
 {
-    public interface IEntityConverter
+    public interface IScalarConverter
     {
         dynamic Convert<T>(DbDataReader reader);
     }
