+++ src/VIC.DataAccess/Core/Converter/DbFuncNameConverter.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using VIC.DataAccess.Abstraction;
+using VIC.DataAccess.Abstraction.Converter;
 
-namespace VIC.DataAccess.Core
+namespace VIC.DataAccess.Core.Converter
 {
     public class DbFuncNameConverter : IDbFuncNameConverter
     {
