+++ src/VIC.ObjectConfig/project.json
@@ -15,7 +15,7 @@
 
   "dependencies": {
     "NETStandard.Library": "1.6.0",
-    "Microsoft.Extensions.Configuration.FileExtensions":"1.0.0"
+    "Microsoft.Extensions.Configuration.FileExtensions": "1.0.0"
   },
 
   "frameworks": {
